Refactor to make specialized curve / surface operations possible
for example, when extruding a Line curve along Vector, you will have a
Plane, not a general Extrusion object.

Move some classes from curve.core and surface.core into curve.primitives
and surface.primitives (others can be moved later if necessary).

utils.surface module does not import utils.surface.algorithms anymore -
that would create cyclic import.

Add a test case that imports all *.py files, to check that there are no
broken / dangling / cyclic imports.
