+++ nodes/surface/revolution_surface.py
@@ -9,7 +9,7 @@ from sverchok.node_tree import SverchCustomTreeNode, throttled
 from sverchok.data_structure import updateNode, zip_long_repeat, ensure_nesting_level
 from sverchok.utils.logging import info, exception
 from sverchok.utils.curve import SvCurve
-from sverchok.utils.surface import SvRevolutionSurface
+from sverchok.utils.surface.algorithms import SvRevolutionSurface
 
 class SvRevolutionSurfaceNode(bpy.types.Node, SverchCustomTreeNode):
     """
