+++ utils/curve/__init__.py
@@ -6,6 +6,7 @@
 # License-Filename: LICENSE
 
 from sverchok.utils.curve.core import *
+from sverchok.utils.curve.primitives import *
 from sverchok.utils.curve.algorithms import *
 from sverchok.utils.curve.bezier import *
 from sverchok.utils.curve.freecad import *
