+++ utils/surface/__init__.py
@@ -8,4 +8,5 @@
 from sverchok.utils.surface.core import *
 from sverchok.utils.surface.data import *
 from sverchok.utils.surface.sphere import *
-from sverchok.utils.surface.algorithms import *
+from sverchok.utils.surface.primitives import *
+
